modularized pins and clients
