+++ app_taifa_flutter/lib/objects/Client.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class Client {
+  String name;
+  Color color;
+
+  Client({
+    required this.name,
+    required this.color,
+  });
+}
