+++ app_taifa_flutter/lib/objects/Pins.dart
@@ -1,0 +1,15 @@
+
+
+class Pins {
+  String name;
+  String client;
+  double latitude;
+  double longitude;
+
+  Pins({
+    required this.name,
+    required this.client,
+    required this.latitude,
+    required this.longitude,
+  });
+}
